Implement classpath indexer.
This commit makes it possible to quickly index a classpath of a project.
This makes it possible to jump to definition from a project source file
to a dependency source file. Note that this does not make is possible to
jump to definition from dependency sources to dependency sources.

While testing this feature, I extended the Jars utility to fetch
-sources jars instead of regular jars and adapted the
scala.meta.Classpath.deep method to read entries of jar files and report
errors using our logging infrastructure.