Revert snapshot tests.
I didn't like how
- the expected output was in a separate file
- it was too easy to create huge snapshot files
I think the ideal snapshot tests should be small and you should be able
to see the expected output alongside the input. This should
be possible by using the same trick as Jane Street does with "expect
tests", where the test framework creates a sibling source file
`X.check`, which is identical to the original test suite `X.scala`
except with the multiline strings fixed. If you are happy with the
expected output in the check file, you can overwrite the original file
with the check file.