Move to re-usable `Endpoint[A, B](method: String)`
The motivaiton for this PR is twofold:

- make it easier to add add/reuse/share protocol handlers/clients
- reorganize the project structure to org.langmeta.{lsp,jsonrpc}
  as the new home for LSP and JSON-RPC specific features.

`Endpoint[A, B](method: String)` is a description of a JSON RPC message.
Endpoints can be used both from the server side to handle messages as
well as to send messages from the client. For example, the completion
endpoint can be defined like this

```scala
val completion: Endpoint[TextDocumentPositionParams, CompletionList] =
  request("textDocument/completion")
```

On the server, this endpoint can be handled with
```
.request(completion) { params => ??? }
```
while on the client it can be requested like this
```scala
implicit val client: LanguageClient = ???
completion.request(TextDocumentPositionParams(???)): Task[CompletionList]
````

Endpoints simplify the large fluid Services.request/notification
statement in ScalametaServices, since it can piggy back on the standard
endpoints avoiding a lot of explicit type parameters and listing out
string method names.

By being first-class values, endpoints enable us to have for example
cleaner interfaces to report messages
```scala
// before
import languageserver.types.MessageType
client.sendMessage(MessageType.Error, String)
// after
sendMessage.error(String)
```

While working on endpoints, I wasn't sure whether to put them in the
languageserver project or metaserver. I took the liberty to reorganize
the package so that:
- languageserver no longer exists, the project has been removed
- org.langmeta.jsonrpc contains JSON-RPC specific types
- org.langmeta.lsp contains LSP specific types

This follows a convention we only recently started trying to follow in
scalameta/sclameta: keep language-agnotic features outside of the
scala.meta package.

While working on this refactoring I hit on the first first debugging
session :( A sloppy copy-paste in `Services.notification` caused handler to be
wrapped in a Task that got converted to Unit and never evaluated.
Beware!