Move to upickle for json.
The transitive dependency graph of upickle (including inlined jawn) is
~500kb while circe is ~8mb (primarily due to 22-tuple generated
implicits in cats). Considering that IntelliJ Scala team might want
to bring in lsp4s we should try to keep size of dependencies to a
minimum.

Quite a few changes were necessary to deliver this migration

- minimal "ping pong" unit test to verify the previous behavior
- `@json` macro annotation to replace `@JsonCodec`
- `Enums.findValues` def macro to replace enumeratum
- re-write all manual decoders/encoders by hand for upickle

While doing the refactoring, I took the liberty to minimize the lsp4s
public API

- Add `final` where possible
- Add `private` where possible
- Remove `case` where possible
- Move as much as possible to `scala.meta.internal` as possible.