Parameter processId to initialize is optional
According to the specification, the processId parameter is optional.

Some LSP clients (for example, vim-lsp) do not provide this parameter.
Currently, lsp4s fails to initialize for those clients. Making this
parameter optional allows initialization to continue.

This parameter appears to be unused, so only the type has been changed.

https://github.com/Microsoft/language-server-protocol/blob/master/versions/protocol-2-x.md#initialize